+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Error and warning status not vocalized for `progress indicator` components (Orange-OpenSource/ouds-ios#1697)
 - Use of asset for typography heading marker (Orange-OpenSource/ouds-ios#1689)
 - Adjust title color on toolbar bar top when content color change (Orange-OpenSource/ouds-ios#1563)
 - Truncated title on `tab bar` focused item when using *Full Keyboard Access* (Orange-OpenSource/ouds-ios#1641)